using a different class for custom parser
