+++ rets/__init__.py
@@ -2,7 +2,7 @@ from .session import Session
 from .exceptions import RETSException
 
 __title__ = 'rets'
-__version__ = '0.4.7'
+__version__ = '0.4.8'
 __author__ = 'REfindly'
 __license__ = 'MIT'
 __copyright__ = 'Copyright 2019 REfindly'
